Improve security for streaming between nodes
